Bug fix and vignette update: - Solved to do related to the functioning of buildSequences as mentioned in the old README version:
Check sequence extract method in buildSequences. If a location descriptor of the INSDC include a start or end location of unknown position (for example, “<345..500”), should the function extract the sequence using the first or end position of the whole sequence instead of the unknown position or include the last/first known position only? (the 345 position as mentioned in the example). As example of a problem arising, try downloading all CDS of the records corresponding to Ophthalmolycus amberensis (specifically the acc num ON417737.1 and the CDS CYTB), or the acc num KX362346.1 for the muts CDS of the species Notisis elongata. Another issue related to this problem can be seen with the accession number “JN034583.1” of the species Dissostichus mawsoni, where different portions of “immunoglobulin M heavy chain” are separated by exons.

- Updated vignette to speed up analyses and README accordingly